Add a auth service with login/logout/isAuthenticaed methods
