+++ src/app/api/api.module.ts
@@ -1,22 +1,24 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { IsAuthenticated } from './guards/is-authenticated';
+import { AuthGuard } from './guards/auth.guard';
 import { SandboxStatusResolver } from './resolvers/sandbox-status-resolver.service';
 import { SandboxesResolver } from './resolvers/sandboxes-resolver.service';
 import { ServicesResolver } from './resolvers/services-resolver.service';
 import { PreferencesStorage } from './services/preferences-storage.service';
 import { DebugStatusApi } from './services/debug-status-api.service';
+import { AuthService } from './services/auth.service';
 
 @NgModule({
   imports: [CommonModule],
   providers: [
-    IsAuthenticated,
+    AuthGuard,
     SandboxStatusResolver,
     SandboxesResolver,
     ServicesResolver,
     PreferencesStorage,
     DebugStatusApi,
+    AuthService,
   ],
 })
 export class ApiModule {}
