+++ src/app/terminal/terminal-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { IsAuthenticated } from '../api/guards/is-authenticated';
+import { AuthGuard } from '../api/guards/auth.guard';
 import { ServicesResolver } from '../api/resolvers/services-resolver.service';
 import { SandboxStatusResolver } from '../api/resolvers/sandbox-status-resolver.service';
 
@@ -10,7 +10,7 @@ import { TerminalViewComponent } from './terminal-view/terminal-view.component';
 const ROUTES: Routes = [
   {
     path: ':sandboxId',
-    canActivate: [IsAuthenticated],
+    canActivate: [AuthGuard],
     component: TerminalViewComponent,
     resolve: {
       status: SandboxStatusResolver,
@@ -22,6 +22,6 @@ const ROUTES: Routes = [
 @NgModule({
   imports: [RouterModule.forChild(ROUTES)],
   exports: [RouterModule],
-  providers: [IsAuthenticated, SandboxStatusResolver, ServicesResolver],
+  providers: [AuthGuard, SandboxStatusResolver, ServicesResolver],
 })
 export class TerminalRoutingModule {}
